requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
langchain==0.1.17
langchain-community==0.0.38
langsmith==0.1.18
langchain-groq==0.1.0
langchain-google-genai==1.0.5
transformers==4.35.0
torch==2.1.0
Pillow==10.1.0
pytesseract==0.3.10
fuzzywuzzy==0.18.0
python-levenshtein==0.25.0
groq==0.4.1
google-generativeai==0.3.2

